requirements.txt
# Production dependencies for AWS deployment
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
flask-limiter==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Werkzeug==2.3.7
cryptography==41.0.4
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
gunicorn==21.2.0

# Document Processing (for in-memory analysis)
PyPDF2==3.0.1
python-docx==1.1.0

# ML/RAG System Dependencies
beautifulsoup4==4.12.2
lxml==5.1.0
numpy==1.25.2
pandas==2.1.3

# Vector Database Options
chromadb>=0.4.22  # Local development only
pinecone>=7.0.0  # Cloud vector DB for production

# Embeddings & LLM
sentence-transformers==2.3.1
# openai==1.6.1  # Uncomment if using OpenAI

# LangChain (optional, for advanced RAG)
# langchain==0.1.0
# langchain-community==0.0.10

# Development dependencies (commented out for production)
# streamlit==1.28.1
# pypdf2==3.0.1
# python-docx==1.1.0
# jinja2==3.1.2
# plotly==5.17.0